[core] #528 Cleanup and using new model classed for index creation jobs. This commit introduces following changes: 1. Removed old logic to parse using the MethodVisitor. 2. Cleanup 3. Written new job to use new model classes. fixes #528 closes #544

+++ core/src/main/scala/com/kodebeagle/indexer/FileMetaDataIndexer.scala
@@ -48,7 +48,7 @@ object FileMetaDataIndexer extends Logger {
 
   // TODO: This class needs to be looked at closely and optimized.
   def generateMetaData(resolver: SingleClassBindingResolver,
-                       unit: CompilationUnit, repoId: Int, fileName: String): FileMetaData = {
+                       unit: CompilationUnit, repoId: Long, fileName: String): FileMetaData = {
     val typesAtPos = resolver.getTypesAtPosition
     // External reference
     val idVsExternalRefs: Map[String, Int] = getExternalRefs(resolver, typesAtPos)
